requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
ultralytics>=8.0.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0

# UniTS official (optional, for full UniTS model)
# timm>=0.9.0
# einops>=0.7.0
